refactor(api): decompose WithingsClient into Facade + 5 Mixins
Split 979-line god class into focused modules following the established
Facade + Mixins pattern used across the codebase (10th god class done).

- CredentialsMixin: load/save/refresh tokens, auth checks
- OAuthMixin: authorization URL generation, code exchange
- HttpMixin: rate limiting, authenticated requests with retry
- SleepMixin: sleep data retrieval, multi-segment aggregation
- MeasurementsMixin: weight/body comp, blood pressure, readiness

Zero breaking change: WithingsClient stays in withings_client.py,
all existing imports and test patches work via MRO.
