+++ magma_cycling/api/withings/__init__.py
@@ -1,0 +1,1 @@
+"""WithingsClient mixin modules for god class decomposition."""
